refactor: [US-047] - Move upload + task-status views into core/views/upload.py
